Split Debt Payoff into its own page
Also fixes a pre-existing bug where the Debt Payoff strategy comparison
and Excel export button were rendered inside the Standard/Lean/Fat FIRE
section instead of the Debt Payoff section.
