+++ app/MyFireNumber/AppShell.xaml.cs
@@ -37,6 +37,7 @@ public partial class AppShell : Shell
 		Routing.RegisterRoute("coast-fire", typeof(CoastFirePage));
 		Routing.RegisterRoute("barista-fire", typeof(BaristaFirePage));
 		Routing.RegisterRoute("reverse-fire", typeof(ReverseFirePage));
+		Routing.RegisterRoute("debt-payoff", typeof(DebtPayoffPage));
 		Routing.RegisterRoute("retirement-annual-details", typeof(RetirementAnnualDetailsPage));
 		Loaded += OnLoaded;
 	}
