+++ app/MyFireNumber/Views/DebtPayoffPage.xaml.cs
@@ -1,0 +1,12 @@
+using MyFireNumber.ViewModels;
+
+namespace MyFireNumber.Views;
+
+public partial class DebtPayoffPage : CalculatorPageBase
+{
+    public DebtPayoffPage(DebtPayoffViewModel viewModel)
+    {
+        InitializeComponent();
+        InitializeCalculator(viewModel);
+    }
+}
